requirements.txt
appdirs==1.4.0
baron==0.6.1
dj-database-url==0.3.0
dj-static==0.0.6
Django==1.8.4
django-activelink==0.4
django-bootstrap-pagination==1.5.1
django-braces==1.8.1
django-cors-headers==1.1.0
django-datetime-widget==0.9.3
django-flat-theme==1.1.1
django-oauth-toolkit==0.9.0
django-sslserver==0.15
django-widget-tweaks==1.4.1
djangorestframework==3.2.4
gunicorn==19.3.0
itsdangerous==0.24
Jinja2==2.8
MarkupSafe==0.23
oauthlib==1.0.3
Pygments==2.0.2
pytz==2015.6
PyYAML==3.11
redbaron==0.5.1
requests==2.7.0
requests-oauthlib==0.5.0
rply==0.7.4
simplejson==3.8.0
six==1.9.0
sqlparse==0.1.16
static3==0.6.1
Werkzeug==0.10.4
wheel==0.26.0
